% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsprintf.R
\name{nsprintf}
\alias{nsprintf}
\title{Easy Internationalized Message Pluralization}
\usage{
nsprintf(n, msg1, msg2, ..., domain = NULL)
}
\arguments{
\item{n}{A numerical value, passed to \code{\link[base]{ngettext}}.}

\item{msg1}{A character string containing an English-language message when 
\code{n == 1}, passed to \code{\link[base]{ngettext}}.}

\item{msg2}{A character string containing an English-language message when 
\code{n != 1}, passed to \code{\link[base]{ngettext}}.}

\item{...}{Additional arguments passed to \code{\link[base]{sprintf}}.}

\item{domain}{A character string specifying a gettext domain, passed to 
\code{\link[base]{ngettext}}.}
}
\value{
A character vector.
}
\description{
A single function to facilitate the correct pluralization and 
internationalization of R messages.
}
\examples{
# simple pluralization
nsprintf(0:3, "there is one apple", "there are \%d apples")

# additional arguments to `sprintf()`
# count the number of objects in each environment on the search path
envs <- search()
n <- vapply(envs, function(env) length(ls(env)), integer(1))
nsprintf(n, "there is \%d variable in \%s", 
            "there are \%d variables in \%s", envs)
            
# nested nsprintf: 
n_objects <- 6
k_chosen <- seq(0, n_objects)
n_combinations <- choose(n_objects, k_chosen)
nsprintf(
  n_combinations, 
  "there is \%d combination of \%s chosen from a set of size \%d", 
  "there are \%d combinations of \%s chosen from a set of size \%d",
  nsprintf(k_chosen, "\%d object", "\%d objects"),  
  n_objects
)
}
\author{
Thomas J. Leeper
}

